Move Razor specific service base
