+++ src/Workspaces/Remote/Razor/Microsoft.CodeAnalysis.Remote.Razor.ServiceHub.csproj
@@ -37,7 +37,7 @@
     <Compile Include="..\ServiceHub\Shared\RoslynJsonConverter.SolutionIdConverters.cs">
       <Link>Shared\RoslynJsonConverter.SolutionIdConverters.cs</Link>
     </Compile>
-    <Compile Include="..\ServiceHub\Shared\ServiceHubServiceBase.cs">
+    <Compile Include="..\ServiceHub\Shared\ServiceBase.cs">
       <Link>Shared\ServiceHubServiceBase.cs</Link>
     </Compile>
     <Compile Include="..\ServiceHub\Shared\RemoteEndPoint.cs">
